+++ src/business/team_members.rs
@@ -40,7 +40,7 @@ pub mod v10 {
     #[derive(Clone, Debug, Default)]
     pub struct ListParams {
         pub created_before: Option<String>,
-        pub limit: Option<u64>,
+        pub limit: Option<u16>,
     }
 
     #[derive(Clone, Debug, Deserialize, Serialize)]
